video_splitter/requirements.txt
PyQt6>=6.0.0
ffmpeg-python>=0.2.0 